mobile_mart_service/requirements.txt
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
psycopg2==2.9.6
